Updated my roles + Added Firebase DB
Please check the discord server. I have put up a simple reaction poll that assigns roles automatically in the server. Since you are all now tagged with your github username as well, that will make the initial designing and first steps of this project easier. 

I advise you assign yourself several roles since in the beginning it may be useful to work on parts we actually need, and later on we can start working more on our individual "expertise".