Added a PaaS Cloud Computing platform for server
Heroku is quite simple to deploy a web app but I think they have limited uptimes, but they are free... in the contrary to Azure. I will try to obtain a license for Azure App Service from a relative but if that fails we will have to use Heroku or find a way to host servers, in favor of PaaS or cloud computing services to avoid security and infrastructure issues.