Added load_dotenv opts + useful enum to config.py
the enum CurrentDatabaseConnection holds CURRENTLY non-security essential database information. YOu are free to change the fields if you want to name your database something else or use a different RDBSM, but to test in such way create a new branch. 

THe info in the enum will change to something more secure and that is hosted once I have setup the final database or when other backend might want to add integration for lets say firebase.